requirements.txt
appdirs==1.4.4
asgiref==3.2.10
distlib==0.3.1
dj-database-url==0.5.0
Django==3.1.1
django-cors-headers==3.5.0
django-crispy-forms==1.9.2
django-rest-auth==0.9.5
django-rest-passwordreset==1.1.0
djangorestframework==3.11.1
filelock==3.0.12
gunicorn==20.0.4
importlib-metadata==1.7.0
psycopg2==2.8.6
pytz==2020.1
six==1.15.0
sqlparse==0.3.1
virtualenv==20.0.31
zipp==3.1.0
